FIN-13 Separating TooltipText and other components. Changing Loading styles according to the design
