+++ src/src_V2/components/index.js
@@ -4,3 +4,4 @@ export { default as Layout } from './layout/Layout';
 export { default as NoData } from './no-data/NoData';
 export { default as Sidebar } from './sidebar/Sidebar';
 export { default as LoadingSpinner } from './loading-spinner/LoadingSpinner';
+export { default as TooltipText } from './tooltip/TooltipText';
