+++ src/src_V2/components/loading-spinner/LoadingSpinner.jsx
@@ -2,10 +2,8 @@ import styles from './LoadingSpinner.module.css';
 
 export default function LoadingSpinner() {
   return (
-    <div className={styles.spinnerDiv}>
-      <div className={styles.spinnerSubDiv}>
-        <div></div>
-      </div>
+    <div className={styles.preloader}>
+      <div className={styles.loader} />
     </div>
   );
 }
