+++ src/src_V2/components/tooltip/TooltipText.module.css
@@ -1,0 +1,6 @@
+.tooltip {
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  max-width: 100px;
+}
